Add files via upload
I think that, in the sake of keeping a simple design, that this is the little bit of tweaking I could do. Centered the title for the form; changed the color to be less contrasting with the chart from at least my current device; and put padding around the content so the webpage did not align with the edges of the browser.